Always test installed version when running `pytest`.
We assume the package has been installed, as it would be during CI runs
or when using Tox. So the modules in the `tests` folder shouldn't
import from the parent folder.

The code in the repository can be tested without installation by
running `python -m pytest` from the root folder. As opposed to
running just `pytest`, this adds the current directory to Python's
search path.